+++ plugin/src/test/kotlin/org/jlleitschuh/gradle/ktlint/UnsupportedGradleTest.kt
@@ -3,7 +3,9 @@ package org.jlleitschuh.gradle.ktlint
 import org.assertj.core.api.Assertions.assertThat
 import org.gradle.util.GradleVersion
 import org.jlleitschuh.gradle.ktlint.testdsl.buildAndFail
+import org.jlleitschuh.gradle.ktlint.testdsl.getMajorJavaVersion
 import org.jlleitschuh.gradle.ktlint.testdsl.project
+import org.junit.jupiter.api.Assumptions
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.condition.DisabledOnOs
@@ -14,7 +16,15 @@ class UnsupportedGradleTest : AbstractPluginTest() {
     @Test
     @DisabledOnOs(OS.WINDOWS)
     internal fun errorOnOldGradleVersion() {
-        project(GradleVersion.version("6.9.2")) {
+        /**
+         * This test ensures the proper error message is printed when an unsupported version of gradle is used.
+         * However, our minimum version of gradle is still 7.x, which will not run at all on Java 21.
+         * Gradle 8.5 is needed for Java 21.
+         * So if java 21 is currently being used, skip this test
+         */
+        Assumptions.assumeFalse(getMajorJavaVersion() >= 21)
+
+        project(GradleVersion.version("7.4.1")) {
             buildAndFail(CHECK_PARENT_TASK_NAME) {
                 assertThat(output).contains(
                     "Current version of plugin supports minimal Gradle version: " +